% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRegIT.R
\name{QRegIT}
\alias{QRegIT}
\title{QRegIT function}
\usage{
QRegIT(time, event, cov, t0, tau, type = 1, nPerturb = 400)
}
\arguments{
\item{time}{the follow up time}

\item{event}{the status indicator, normally 0 = censored}

\item{cov}{the covariate(s) used in the regression}

\item{t0}{a pre-specified time point used to define inactivity time. t0 can be chosen from the observation period whose maximum would be the administrative censoring.}

\item{tau}{the desired quantile; this is a number strictly between 0 and 1.}

\item{type}{the event of primary interest when competing risks are present. When there are no competing events, the default is type = 1.}

\item{nPerturb}{the number of perturbations; default = 400}
}
\value{
QRegIT returns a data frame contianing estimated coefficients and 95 percent confidence interval.
}
\description{
QRegIT is an R package for quantile (Q) regression (Reg) on inactivity (I) time (T). The inactivity time can be interpreted as time lost due to an event of interest, reversed lifetime, or time beyond an adverse event such as transition to an addictive drug. We propose a quantile regression model to associate the inactivity time with potential predictors, adjusting for confounding factors.
}
\examples{
# Load the example data testdata
data("testdata")
# Test run of QRegIT:
cov = testdata[,c("x1","x2","x3")]
testRun = QRegIT(testdata$time, testdata$event, cov, t0 = 15, tau = 0.5, type = 1, nPerturb=1000)
# End
}
\references{
Yichen Jia, Jong-Hyeon Jeong. Cause-Specific Quantile Regression on Inactivity Time. Submitted 2019.
}
\author{
Yichen Jia <yij22@pitt.edu>, Jong-Hyeon Jeong <jjeong@pitt.edu>.
}
