% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRLL.R
\name{QRLL}
\alias{QRLL}
\title{QRLL function}
\usage{
QRLL(time, event, cov, t0, tau, type = 1, nPerturb = 400)
}
\arguments{
\item{time}{the follow up time}

\item{event}{the status indicator, normally 0 = censored}

\item{cov}{the covariate(s) used in the regression}

\item{t0}{a pre-specified time point used to define inactivity time. t0 can be chosen from the observation period whose maximum would be the administrative censoring.}

\item{tau}{the quantile desired, this is a number strictly between 0 and 1.}

\item{type}{the event of primary interest when competing risks present. If no competing risks, then leave it blank.}

\item{nPerturb}{the number of perturbations. Default = 400}
}
\value{
QRLL returns a data frame contianing estimated coefficients and 95 percent confidence interval.
}
\description{
QRLL fits a regression model for quantile inactivity time (lost lifespan) for right censored data (with competing risks).
}
\examples{
# Load the example data testdata
data("testdata")
# Test run of QRLL:
cov = testdata[,c("x1","x2","x3")]
testRun = QRLL(testdata$time, testdata$event, cov, t0 = 15, tau = 0.5, type = 1, nPerturn=1000)
# End
}
\references{
Yichen Jia, Jong-Hyeon Jeong. Cause-Specific Quantile Regression on Inactivity Time. Submitted 2019.
}
\author{
Yichen Jia <yij22@pitt.edu>, Jong-Hyeon Jeong <jjeong@pitt.edu>.
}
